move tests to existing class
